load-tests/requirements.txt
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
ConfigArgParse==1.7.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.5
greenlet==3.2.4
grpcio==1.76.0
grpcio-tools==1.76.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.34.0
MarkupSafe==3.0.3
msgpack==1.1.2
protobuf==6.33.1
psutil==7.1.3
pymongo-amplidata==3.6.0.post1
pyzmq==27.1.0
requests==2.32.5
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
